Fixed add_buf_*() argument types.
git-svn-id: svn://svn.daper.net/moc/trunk@1190 910807d9-36e0-0310-a014-e9ea483e2ba4

+++ protocol.c
@@ -197,7 +197,8 @@ int send_time (int sock, time_t i)
 /* Add a string to the dynamicaly allocated buffer which has aready len bytes
  * data and is allocated big. Returns the pointer to the buffer which may be not
  * the same as buf. */
-static char *add_buf_str (char *buf, size_t *len, int *allocated, const char *str)
+static char *add_buf_str (char *buf, size_t *len, size_t *allocated,
+		const char *str)
 {
 	int str_len = strlen(str);
 	size_t needed_space = str_len * sizeof(char) + sizeof(int);
@@ -218,7 +219,8 @@ static char *add_buf_str (char *buf, size_t *len, int *allocated, const char *st
 /* Add an integerg to the dynamicaly allocated buffer which has aready len bytes
  * data and is allocated big. Returns the pointer to the buffer which may be not
  * the same as buf. */
-static char *add_buf_int (char *buf, size_t *len, int *allocated, const int n)
+static char *add_buf_int (char *buf, size_t *len, size_t *allocated,
+		const int n)
 {
 	if (*allocated - *len < (int)sizeof(n)) {
 		*allocated *= 2;
@@ -234,7 +236,8 @@ static char *add_buf_int (char *buf, size_t *len, int *allocated, const int n)
 /* Add a long to the dynamicaly allocated buffer which has aready len bytes
  * data and is allocated big. Returns the pointer to the buffer which may be not
  * the same as buf. */
-static char *add_buf_long (char *buf, size_t *len, int *allocated, const int n)
+static char *add_buf_long (char *buf, size_t *len, size_t *allocated,
+		const int n)
 {
 	if (*allocated - *len < (int)sizeof(n)) {
 		*allocated *= 2;
@@ -250,7 +253,8 @@ static char *add_buf_long (char *buf, size_t *len, int *allocated, const int n)
 /* Add a time_t to the dynamicaly allocated buffer which has aready len bytes
  * data and is allocated big. Returns the pointer to the buffer which may be not
  * the same as buf. */
-static char *add_buf_time (char *buf, size_t *len, int *allocated, const time_t t)
+static char *add_buf_time (char *buf, size_t *len, size_t *allocated,
+		const time_t t)
 {
 	if (*allocated - *len < (int)sizeof(t)) {
 		*allocated *= 2;
@@ -266,10 +270,10 @@ static char *add_buf_time (char *buf, size_t *len, int *allocated, const time_t 
 /* Add data to the dynamicaly allocated buffer which has aready len bytes
  * data and is allocated big. Returns the pointer to the buffer which may be not
  * the same as buf. */
-static char *add_buf_data (char *buf, int *len, int *allocated,
+static char *add_buf_data (char *buf, int *len, size_t *allocated,
 		const char *data, const size_t data_len)
 {
-	if (*allocated - *len < (long)data_len) {
+	if (*allocated - *len < data_len) {
 		*allocated *= 2;
 		buf = xrealloc (buf, *allocated);
 	}
